+++ src/main/java/autopartner/controller/rest/TaskTypeController.java
@@ -31,7 +31,7 @@ public class TaskTypeController {
 
     @Secured({"ROLE_ADMIN", "ROLE_ROOT"})
     @RequestMapping(value = "/api/task/type/{id}", method = RequestMethod.GET)
-    public ResponseEntity<TaskType> get(@PathVariable Integer id) {
+    public ResponseEntity<TaskType> get(@PathVariable Long id) {
         return ResponseEntity.ok(taskTypeService.getTaskTypeById(id));
     }
