+++ src/main/java/autopartner/repository/CarBrandRepository.java
@@ -4,6 +4,6 @@ import autopartner.domain.entity.CarBrand;
 import autopartner.domain.entity.CarType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CarBrandRepository extends CrudRepository<CarBrand, Integer> {
+public interface CarBrandRepository extends CrudRepository<CarBrand, Long> {
     Iterable<CarBrand> findByActiveTrue();
 }
