+++ src/main/java/autopartner/repository/MaterialRepository.java
@@ -3,6 +3,6 @@ package autopartner.repository;
 import autopartner.domain.entity.Material;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MaterialRepository extends CrudRepository<Material, Integer> {
+public interface MaterialRepository extends CrudRepository<Material, Long> {
     Iterable<Material> findByActiveTrue();
 }
